files of the functional makeblock-drive library
I am adding all files of the makeblock-drive library that works with the mbot, to try and find the difference causing the issue

+++ src/MeEncoderOnBoard.h
@@ -50,7 +50,7 @@
  *    23. void MeEncoderOnBoard::setSpeedPid(float p,float i,float d);
  *    24. void MeEncoderOnBoard::setPosPid(float p,float i,float d);
  *    25. void MeEncoderOnBoard::setPulse(int16_t pulseValue);
- *    26. void MeEncoderOnBoard::setRatio(float RatioValue);
+ *    26. void MeEncoderOnBoard::setRatio(int16_t RatioValue);
  *    27. void MeEncoderOnBoard::setMotionMode(int16_t motionMode);
  *    28. int16_t MeEncoderOnBoard::pidPositionToPwm(void);
  *    29. int16_t MeEncoderOnBoard::speedWithoutPos(void);
@@ -590,7 +590,7 @@ public:
  * \par Others
  *    None
  */
-  void setRatio(float RatioValue);
+  void setRatio(int16_t RatioValue);
 
 /**
  * \par Function
@@ -719,8 +719,8 @@ private:
    uint8_t _IntNum;
    uint8_t _Slot;
    int16_t _Encoder_output;
-   unsigned long _Measurement_speed_time;
-   unsigned long _Encoder_move_time;
+   long _Measurement_speed_time;
+   long _Encoder_move_time;
    bool _enabled;
    cb _callback;
 };
